pyproject.toml
[project]
name = "flask-email-system"
version = "0.1.0"
description = "A learning project to build a basic email system with Flask, featuring user authentication, JWT, MongoDB, password hashing, and encryption."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "cryptography>=45.0.4",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-limiter>=3.12",
    "flask-mail>=0.10.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
]
